tidyWrapper.pl: use Perl::Tidy::Sweetened if available
Perl::Tidy::Sweetened adds support for some syntactic sugar provided by
some modules. The module uses Perl::Tidy's prefilter and postfilter hooks
to support `method` and `func` keywords (which are also supported on more
recent versions of Perl::Tidy) and other stuff.

tidyWrapper.pl is now additionally checking for the availability of the
addon module and switches out the `perltidy` method that's being used
for tidying.

We are using the builtin universal `can` method to get references to the
given sub so we can call them easily later.

More details:
https://metacpan.org/pod/Perl::Tidy::Sweet
https://perldoc.perl.org/UNIVERSAL#CLASS-%3Ecan(-METHOD-)