Fix warnings via namespace::autoclean
On Perl 5.40 a new warning is introduced when the import method of an
unknown package is called.

https://perldoc.perl.org/perldelta#Calling-the-import-method-of-an-unknown-package-produces-a-warning

This is seen via PerlNavigator because Inquisitor.pm removes
namespace::clean and namespace::autoclean from %INC. The solution is to
add stub import subs for these packages so that the Perl interpreter no
longer believes that they do not exist.