server: pod: rework symbol lookup
This commit adds support for the following:
- symbol lookup for `=item` paragraphs, restoring the original
  behaviour
- consecutive `=item` or  `=head\d` paragraphs (of the same level) for
  the same subroutine, method, etc., e.g. in order to show an
  alternative way to call the subroutine, are now displayed as
  expected
- support symbol lookup via index entries (`X<entry>`), which may be
  used to e.g. look up an alias of a subroutine

To elaborate on the last point, here's a quick example:

`File::Copy::move` has an alias called `File::Copy::mv`. The
documentation for move contains an index entry for `mv`, so that the
docs count for both subroutines:

  =item move
  X<move> X<mv> X<rename>

The symbol lookup regex will now also match index entries, so if a
signature doesn't match, the regex will try to match for an index
entry, if there is any.

This makes it possible to look up the docstring for `File::Copy::mv`.

Also, add documentation for the `lookupSymbolInPod` function.

Signed-off-by: Max R. Carrara <max@aequito.sh>