server: pod: fix symbol lookup edge case regarding list items
Some Perl modules in the wild document their subroutines like this:

  =over

  =item *

  $obj->foo()

  Lorem ipsum odor amet, consectetuer adipiscing elit. Quis a habitasse
  hendrerit efficitur phasellus lacus.

  [...]

  =back

While this would render to Markdown as expected, the =item paragraph
would still contain no text in terms of how the data is represented.

This commit makes it so that if an =item without text is followed by
an ordinary paragraph, the latter is melded into the former during the
initial parse of the POD document. Thus, symbol lookup for this case
in particular works again as expected.

Test cases are adapted correspondingly.

Signed-off-by: Max R. Carrara <max@aequito.sh>