server: pod: remove error handling of RawPodParser
As suggested in PR #142 [1], get rid of the error handling in the
RawPodParser class as part of making the overall POD-to-Markdown
conversion process more lax.

This only affects =begin, =end and =for command paragraphs, which are
now parsed even if there is no format name, despite the spec [2]
requiring one.

[1]: https://github.com/bscan/PerlNavigator/pull/142#issuecomment-2568327657
[2]: https://perldoc.perl.org/perlpodspec#%22=begin-formatname%22

Signed-off-by: Max R. Carrara <max@aequito.sh>